+++ .changeset/sandbox-serverless-runtime.md
@@ -16,5 +16,6 @@ Serverless/edge run model for the sandbox layer — a trigger can start an agent
 - **Claude Code runs on the Cloudflare sandbox** — a new `SandboxCapabilities.writableStdin` flag lets a provider advertise that spawned processes have no writable host→process stdin (Cloudflare). The Claude Code adapter detects this and delivers the prompt via a file + in-shell stdin-redirection (`claude -p … < file`) instead of a host stdin write, keeping it out of argv. Host/Docker are unchanged (stdin path).
 - **Tool-bridge token hardening** — the per-run bearer token is no longer passed inline in argv. The Claude Code adapter writes the bridge MCP config to a file and passes claude the path (`--mcp-config <file>`), so the token can't be read from `ps` / `/proc/<pid>/cmdline` by other processes in the sandbox.
 - **Co-located ("combined") run model** — new `@tanstack/ai-sandbox` exports (`remoteToolStubs`, `toolDescriptors`, `httpRemoteToolExecutor`, `executeHostTool`, plus the `RemoteToolExecutor` interface) let the harness loop AND its MCP tool-bridge run INSIDE the container (the in-container sandbox is just `local-process`, with native stdin + a localhost `node:http` bridge). The orchestrator serializes its tools with `toolDescriptors`; the container rebuilds them as delegating stubs with `remoteToolStubs(descriptors, httpRemoteToolExecutor(url, token))`; the orchestrator answers that one call with `executeHostTool(tools, name, args)`. Only host-tool **execution** crosses the container→orchestrator boundary — the MCP transport itself never leaves the container, shrinking the public surface from the whole MCP protocol to a single authenticated tool-exec call.
+- **`@tanstack/ai-sandbox-cloudflare` ships the runtime, not just the example** — the Worker router, the coordinator Durable Object, the durable run-log, and the `POST /run` wire contract are now package code, exported from two entries: `@tanstack/ai-sandbox-cloudflare/agent` (Workers-only — `createCloudflareSandboxAgent`, the coordinator base classes, `DurableObjectRunEventLog`, plus the shared `ContainerRunRequest` + `parseContainerRunRequest`), and `@tanstack/ai-sandbox-cloudflare/runner` (Node — `runInContainerHarness`, the in-container `chat()` server that validates the request, builds `chat()` over `localProcessSandbox()`, and streams NDJSON back). An app supplies only its config + an adapter resolver: the co-located example's whole worker + container program collapse to ~14 lines. Adds `@tanstack/ai-sandbox-local-process` as a peer dependency (used by the `/runner` entry).
 
-See `examples/sandbox-cloudflare-agent` for the Worker → Durable Object → Container reference (trigger returns immediately; the DO coordinates, persists to a DO-backed `RunEventLog`, serves the bridge from its `fetch` handler, and streams to clients over a hibernatable WebSocket), and `examples/sandbox-cloudflare-agent-colocated` for the co-located variant (harness + bridge in the container; only `executeHostTool` crosses back).
+See `examples/sandbox-cloudflare-agent` for the Worker → Durable Object → Container reference (trigger returns immediately; the DO coordinates, persists to a DO-backed `RunEventLog`, serves the bridge from its `fetch` handler, and streams to clients over a hibernatable WebSocket), and `examples/sandbox-cloudflare-agent-colocated` for the co-located variant (harness + bridge in the container; only `executeHostTool` crosses back) — now a one-`createCloudflareSandboxAgent`-call worker plus a one-`runInContainerHarness`-call container program.
