+++ examples/sandbox-cloudflare-agent-colocated/Dockerfile
@@ -2,9 +2,10 @@
 #
 # Unlike examples/sandbox-cloudflare-agent (whose container only runs the
 # `claude` CLI, driven from the DO), THIS image also ships the in-container
-# HARNESS RUNNER (src/container-runner.ts, bundled). That runner is what runs
-# `chat()` + the MCP tool-bridge ON the container's own localhost — the whole
-# point of the co-located model.
+# HARNESS RUNNER (src/container-runner.ts, bundled). That ~6-line program just
+# calls `runInContainerHarness()` from `@tanstack/ai-sandbox-cloudflare/runner`,
+# which runs `chat()` + the MCP tool-bridge ON the container's own localhost —
+# the whole point of the co-located model.
 #
 # It extends the official @cloudflare/sandbox base image (which ships the
 # in-container CONTROL server on port 3000 — exec/files/processes/ports — that
@@ -19,9 +20,11 @@ RUN npm install -g @anthropic-ai/claude-code && claude --version
 # The BUNDLED in-container runner. Build it first on the host with:
 #   pnpm build:runner      # esbuild → dist/container-runner.mjs (single ESM file)
 # It is bundled (not `npm install`ed) so the image stays self-contained and the
-# workspace `@tanstack/*` deps are inlined. The DO starts it via the sandbox
-# control server (`startProcess "node /app/container-runner.mjs"`), then reaches
-# it on RUNNER_PORT with containerFetch — see src/coordinator.ts ensureRunner().
+# workspace `@tanstack/*` deps (incl. the package's `/runner` entry) are inlined.
+# The DO starts it via the sandbox control server
+# (`startProcess "node /app/container-runner.mjs"`), then reaches it on
+# RUNNER_PORT with containerFetch — see the package's ContainerSandboxCoordinator
+# (`@tanstack/ai-sandbox-cloudflare/agent`) ensureRunner().
 COPY dist/container-runner.mjs /app/container-runner.mjs
 
 # IN-CONTAINER LOCAL-PROCESS HAS WRITABLE STDIN. Because the runner uses
