+++ examples/sandbox-cloudflare-agent-colocated/package.json
@@ -13,12 +13,13 @@
     "@tanstack/ai": "workspace:*",
     "@tanstack/ai-claude-code": "workspace:*",
     "@tanstack/ai-sandbox": "workspace:*",
+    "@tanstack/ai-sandbox-cloudflare": "workspace:*",
     "@tanstack/ai-sandbox-local-process": "workspace:*"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20260317.1",
-    "@types/node": "^24.10.3",
-    "esbuild": "^0.25.0",
+    "@types/node": "^24.10.1",
+    "esbuild": "^0.25.12",
     "wrangler": "^4.88.0"
   }
 }
