+++ examples/sandbox-cloudflare-agent-colocated/tsconfig.runner.json
@@ -1,8 +1,9 @@
 {
-  // The in-container harness runner. A plain Node program (node:http server)
-  // bundled into the container image with esbuild — type-checked against
-  // @types/node, NOT @cloudflare/workers-types.
-  "include": ["src/container-runner.ts", "src/protocol.ts"],
+  // The in-container harness runner. A plain Node program bundled into the
+  // container image with esbuild — type-checked against @types/node, NOT
+  // @cloudflare/workers-types. The server + chat() wiring now ships from the
+  // package's `/runner` entry; this app supplies only the adapter.
+  "include": ["src/container-runner.ts"],
   "compilerOptions": {
     "target": "ES2022",
     "module": "ESNext",
