+++ packages/ai-sandbox-cloudflare/package.json
@@ -26,6 +26,14 @@
     ".": {
       "types": "./dist/esm/index.d.ts",
       "import": "./dist/esm/index.js"
+    },
+    "./agent": {
+      "types": "./dist/esm/agent.d.ts",
+      "import": "./dist/esm/agent.js"
+    },
+    "./runner": {
+      "types": "./dist/esm/runner.d.ts",
+      "import": "./dist/esm/runner.js"
     }
   },
   "files": [
@@ -46,11 +54,16 @@
     "@cloudflare/sandbox": "^0.12.1"
   },
   "peerDependencies": {
-    "@tanstack/ai-sandbox": "workspace:^"
+    "@tanstack/ai": "workspace:^",
+    "@tanstack/ai-sandbox": "workspace:^",
+    "@tanstack/ai-sandbox-local-process": "workspace:^"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20241230.0",
+    "@cloudflare/workers-types": "^4.20260317.1",
+    "@tanstack/ai": "workspace:*",
     "@tanstack/ai-sandbox": "workspace:*",
+    "@tanstack/ai-sandbox-local-process": "workspace:*",
+    "@types/node": "^24.10.1",
     "@vitest/coverage-v8": "4.0.14"
   }
 }
